Reorganized and added library of events
